updated to websocket_manager
